\name{gbm_1}
\alias{gbm_1}
\docType{data}
\title{
GBM model
}
\description{
contains trained weight of GBM model
}
\usage{data("gbm_1")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(gbm_1)
## predict(iCTC::gbm_1,iCTC::Original_data_1$Clearcell_Polaris_sample_test_1)
}
\keyword{datasets}
