\name{gbm_2}
\alias{gbm_2}
\docType{data}
\title{
GBM model
}
\description{
contains trained weight of GBM model
}
\usage{data("gbm_2")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(gbm_2)
## predict(iCTC::gbm_2,iCTC::PCA_data_2$Clearcell_Polaris_sample_test_2)
}
\keyword{datasets}
