\name{gbm_3}
\alias{gbm_3}
\docType{data}
\title{
GBM model
}
\description{
contains trained weight of GBM model
}
\usage{data("gbm_3")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(gbm_3)
## predict(iCTC::gbm_3,iCTC::harmony_data_3$Clearcell_Polaris_sample_test_3)
}
\keyword{datasets}
