\name{harmony_data_3}
\alias{harmony_data_3}
\docType{data}
\title{
Training Data
}
\description{
batch corrected and intigrated data using harmony
}
\usage{data("harmony_data_3")}
\format{
  The format is: list of 6 matrices
}
\details{
batch corrected and intigrated data using harmony after PCA (with 100 principle components)
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(harmony_data_3)
## summary(iCTC::harmony_data_3)
}
\keyword{datasets}
