\name{nb_1}
\alias{nb_1}
\docType{data}
\title{
NB model
}
\description{
contains trained weight of NB model
}
\usage{data("nb_1")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(nb_1)
## predict(iCTC::nb_1,iCTC::Original_data_1$Clearcell_Polaris_sample_test_1)
}
\keyword{datasets}
