\name{nb_2}
\alias{nb_2}
\docType{data}
\title{
NB model
}
\description{
contains trained weight of NB model
}
\usage{data("nb_2")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(nb_2)
## predict(iCTC::nb_2,iCTC::PCA_data_2$Clearcell_Polaris_sample_test_2)
}
\keyword{datasets}
