\name{nb_3}
\alias{nb_3}
\docType{data}
\title{
NB model
}
\description{
contains trained weight of NB model
}
\usage{data("nb_3")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(nb_3)
## predict(iCTC::nb_3,iCTC::harmony_data_3$Clearcell_Polaris_sample_test_3)
}
\keyword{datasets}
