\name{Original_data_1}
\alias{Original_data_1}
\docType{data}
\title{
Training Data
}
\description{
filtered and normalized data
}
\usage{data("Original_data_1")}
\format{
  The format is: list of 6 matrices
}
\details{
filtered and normalized  data
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(Original_data_1)
## summary(iCTC::Original_data_1)
}
\keyword{datasets}
