\name{pca_data_100}
\alias{pca_data_100}
\docType{data}
\title{
PCA model
}
\description{
contains trained weight of PCA
}
\usage{data("pca_data_100")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(nb_1)
## predict(iCTC::pca_data_100,iCTC::Original_data_1$Clearcell_Polaris_sample_test_1)
}
\keyword{datasets}
