\name{PCA_data_2}
\alias{PCA_data_2}
\docType{data}
\title{
Training Data
}
\description{
PCA
}
\usage{data("PCA_data_2")}
\format{
  The format is: list of 6 matrices
}
\details{
PCA (with 100 principle components)
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(PCA_data_2)
## summary(iCTC::PCA_data_2)
}
\keyword{datasets}
