\name{raw_test_data}
\alias{raw_test_data}
\docType{data}
\title{
Training Data
}
\description{
raw data
}
\usage{data("raw_test_data")}
\format{
  The format is: list of 5 matrices for testing
}
\details{
raw count data
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(raw_test_data)
## summary(iCTC::raw_test_data)
}
\keyword{datasets}
