\name{rf_1}
\alias{rf_1}
\docType{data}
\title{
RF model
}
\description{
contains trained weight of RF model
}
\usage{data("rf_1")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(nb_1)
## predict(iCTC::rf_1,iCTC::Original_data_1$Clearcell_Polaris_sample_test_1)
}
\keyword{datasets}
