\name{rf_2}
\alias{rf_2}
\docType{data}
\title{
RF model
}
\description{
contains trained weight of RF model
}
\usage{data("rf_2")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(nb_2)
## predict(iCTC::rf_2,iCTC::PCA_data_2$Clearcell_Polaris_sample_test_2)
}
\keyword{datasets}
