\name{rf_3}
\alias{rf_3}
\docType{data}
\title{
RF model
}
\description{
contains trained weight of RF model
}
\usage{data("rf_3")}
\format{
  A object to save trained weight
}
\source{
iCTC
}
\references{
iCTC
}
\examples{
data(nb_3)
## predict(iCTC::rf_3,iCTC::harmony_data_3$Clearcell_Polaris_sample_test_3)
}
\keyword{datasets}
